(PC-10043) Add ImageTile to ui components
